// src/app/models/usuarios.model.ts
export class Usuarios{
    inicio = false;
    adUserId: number | undefined;
    email: string | undefined;
    pwd: string | undefined;
    adId: string | undefined;
    firtsName: string | undefined;
    lastName: string | undefined;
    userName: string | undefined;
    position: string | undefined;
    area: string | undefined;
}