// src/app/models/usuarioRolesLast.model.ts
export class UsuarioRolesLast{
    userRolId: number | undefined;
    userId: number | undefined;
    userADId: string | undefined;
    userName: string | undefined;
    funcionalityId: number | undefined;
    codFunc: string | undefined;
    nameFunc: string | undefined;
    rolId: number | undefined;
    rolName: string | undefined;
}